backend/pyproject.toml
[project]
name = "hawkeye-backend"
version = "0.1.0"
description = "HAWKEYE insider-threat detection backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "sqlalchemy>=2.0.30",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "redis[hiredis]>=5.0.4",
    "confluent-kafka>=2.4.0",
    "neo4j>=5.20.0",
    "lightgbm>=4.3.0",
    "shap>=0.45.1",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "pyarrow>=16.0.0",
    "langchain>=0.2.0",
    "langchain-anthropic>=0.1.15",
    "langchain-community>=0.2.0",
    "anthropic>=0.28.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
    "tenacity>=8.3.0",
    "structlog>=24.2.0",
    "prometheus-client>=0.20.0",
    "minio>=7.2.7",
    "mlflow>=2.13.0",
    "sentry-sdk[fastapi]>=2.5.0",
    "python-multipart>=0.0.9",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.8",
    "mypy>=1.10.0",
    "types-redis>=4.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
